+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Stop calling `camera.refresh()` as that is unnecessary since `v1.2.2`
 - Avoid empty lasso events
 - Fix incorrectly initialized `pointConnectionColorBy`, `pointConnectionOpacityBy`, and `pointConnectionSizeBy`
+- Fix an issue with the color conversion to RGBA
 
 ## v0.16.1